Extract Validate, Convert, and Compare API/App tests
