+++ src/tests/api/test_api.py
@@ -22,12 +22,7 @@ from oauthlib.common import generate_token
 from requests import get
 from rest_framework.test import APITestCase
 
-from api.models import (
-    CompareFileUpload,
-    ConvertFileUpload,
-    SubmitLicenseModel,
-    ValidateFileUpload,
-)
+from api.models import SubmitLicenseModel
 from api.oauth import generate_github_access_token, get_user_from_token
 from api.views import generateLicenseXml
 from src.secret import getAuthCode, getGithubKey, getGithubSecret, getRedisHost
@@ -44,388 +39,6 @@ def isRedisAvailable():
         return True
     except (redis_lib.exceptions.ConnectionError, Exception):
         return False
-
-
-class ValidateFileUploadTests(APITestCase):
-    """Test for validate api with all
-    possible combination of POST and GET
-    request with login enabled.
-    """
-
-    def setUp(self):
-        self.username = "validateapitestuser"
-        self.password = "validateapitestpass"
-        self.tearDown()
-        self.credentials = {"username": self.username, "password": self.password}
-        u = User.objects.create_user(**self.credentials)
-        u.is_staff = True
-        u.save()
-        self.tv_file = open(getExamplePath("SPDXTagExample-v2.0.spdx"))
-        self.rdf_file = open(getExamplePath("SPDXRdfExample-v2.2.spdx.rdf"))
-        self.invalid_tv_file = open(getExamplePath("SPDXTagExample-v2.0_invalid.spdx"))
-        self.invalid_rdf_file = open(getExamplePath("SPDXRdfExample-v2.0_invalid.rdf"))
-
-    def tearDown(self):
-        try:
-            u = User.objects.get_by_natural_key(self.username)
-            u.delete()
-        except ObjectDoesNotExist:
-            pass
-        ValidateFileUpload.objects.all().delete()
-
-    def test_validate_api(self):
-        """Access get without login"""
-        resp1 = self.client.get(reverse("validate-api"))
-        self.assertTrue(resp1.status_code, 403)
-        self.client.login(username=self.username, password=self.password)
-        """ Access get after login"""
-        resp2 = self.client.get(reverse("validate-api"))
-        self.assertTrue(resp2.status_code, 200)
-        """ Valid Tag Value File"""
-        resp3 = self.client.post(
-            reverse("validate-api"),
-            {"file": self.tv_file, "format": "TAG"},
-            format="multipart",
-        )
-        self.assertEqual(resp3.status_code, 200)
-        self.assertEqual(
-            resp3.data["owner"], User.objects.get_by_natural_key(self.username).id
-        )
-        self.assertEqual(resp3.data["result"], "This SPDX document is valid.")
-        """ Valid RDF File"""
-        resp4 = self.client.post(
-            reverse("validate-api"),
-            {"file": self.rdf_file, "format": "RDFXML"},
-            format="multipart",
-        )
-        self.assertEqual(resp4.status_code, 200)
-        self.assertEqual(
-            resp4.data["owner"], User.objects.get_by_natural_key(self.username).id
-        )
-        self.assertEqual(resp4.data["result"], "This SPDX document is valid.")
-        """ Invalid Tag Value File"""
-        resp5 = self.client.post(
-            reverse("validate-api"),
-            {"file": self.invalid_tv_file, "format": "TAG"},
-            format="multipart",
-        )
-        self.assertEqual(
-            resp5.data["owner"], User.objects.get_by_natural_key(self.username).id
-        )
-        self.assertEqual(resp5.status_code, 400)
-        self.assertNotEqual(resp5.data["result"], "This SPDX document is valid.")
-        """ Invalid RDF File"""
-        resp6 = self.client.post(
-            reverse("validate-api"),
-            {"file": self.invalid_rdf_file, "format": "RDFXML"},
-            format="multipart",
-        )
-        self.assertEqual(
-            resp6.data["owner"], User.objects.get_by_natural_key(self.username).id
-        )
-        self.assertEqual(resp6.status_code, 400)
-        self.assertNotEqual(resp6.data["result"], "This SPDX document is valid.")
-        self.client.logout()
-        self.tearDown()
-
-    def test_validate_without_argument(self):
-        self.client.login(username=self.username, password=self.password)
-        resp7 = self.client.post(reverse("validate-api"), {}, format="multipart")
-        self.assertEqual(resp7.status_code, 400)
-        self.client.logout()
-        self.tearDown()
-
-
-class ConvertFileUploadTests(APITestCase):
-    """Test for convert api with all
-    possible combination of POST and GET
-    request with login enabled.
-    """
-
-    def setUp(self):
-        self.username = "convertapitestuser"
-        self.password = "convertapitestpass"
-        self.tearDown()
-        self.credentials = {"username": self.username, "password": self.password}
-        u = User.objects.create_user(**self.credentials)
-        u.is_staff = True
-        u.save()
-        self.tag = "TAG"
-        self.rdf = "RDFXML"
-        self.xlsx = "XLS"
-        self.tv_file = open(getExamplePath("SPDXTagExample-v2.2.spdx"), "rb")
-        self.rdf_file = open(getExamplePath("SPDXRdfExample-v2.2.spdx.rdf"), "rb")
-        self.xlsx_file = open(getExamplePath("SPDXSpreadsheetExample-v2.2.xls"), "rb")
-
-    def tearDown(self):
-        try:
-            u = User.objects.get_by_natural_key(self.username)
-            u.delete()
-        except ObjectDoesNotExist:
-            pass
-        ConvertFileUpload.objects.all().delete()
-
-    def test_convert_api(self):
-        """Access get without login"""
-        resp1 = self.client.get(reverse("convert-api"))
-        self.assertTrue(resp1.status_code, 403)
-        """Access get after login"""
-        self.client.login(username=self.username, password=self.password)
-        resp2 = self.client.get(reverse("convert-api"))
-        self.assertTrue(resp2.status_code, 200)
-        self.client.logout()
-
-    def test_convert_tagtordf_api(self):
-        self.client.login(username=self.username, password=self.password)
-        resp = self.client.post(
-            reverse("convert-api"),
-            {
-                "file": self.tv_file,
-                "from_format": self.tag,
-                "to_format": self.rdf,
-                "cfilename": "tagtordf-apitest",
-            },
-            format="multipart",
-        )
-        self.assertEqual(200, resp.status_code)
-        self.assertEqual(resp.data["message"], "Success")
-        self.assertTrue(resp.data["result"].startswith(settings.MEDIA_URL))
-        self.assertEqual(
-            resp.data["owner"], User.objects.get_by_natural_key(self.username).id
-        )
-        self.client.logout()
-
-    def test_convert_tagtoxlsx_api(self):
-        self.client.login(username=self.username, password=self.password)
-        resp = self.client.post(
-            reverse("convert-api"),
-            {
-                "file": self.tv_file,
-                "from_format": self.tag,
-                "to_format": self.xlsx,
-                "cfilename": "tagtoxlsx-apitest",
-            },
-            format="multipart",
-        )
-        self.assertEqual(200, resp.status_code)
-        self.assertEqual(resp.data["message"], "Success")
-        self.assertTrue(resp.data["result"].startswith(settings.MEDIA_URL))
-        self.assertEqual(
-            resp.data["owner"], User.objects.get_by_natural_key(self.username).id
-        )
-        self.client.logout()
-
-    def test_convert_rdftotag_api(self):
-        self.client.login(username=self.username, password=self.password)
-        resp = self.client.post(
-            reverse("convert-api"),
-            {
-                "file": self.rdf_file,
-                "from_format": self.rdf,
-                "to_format": self.tag,
-                "cfilename": "rdftotag-apitest",
-            },
-            format="multipart",
-        )
-        self.assertEqual(200, resp.status_code)
-        self.assertEqual(resp.data["message"], "Success")
-        self.assertTrue(resp.data["result"].startswith(settings.MEDIA_URL))
-        self.assertEqual(
-            resp.data["owner"], User.objects.get_by_natural_key(self.username).id
-        )
-        self.client.logout()
-
-    def test_convert_rdftoxlsx_api(self):
-        self.client.login(username=self.username, password=self.password)
-        resp = self.client.post(
-            reverse("convert-api"),
-            {
-                "file": self.rdf_file,
-                "from_format": self.rdf,
-                "to_format": self.xlsx,
-                "cfilename": "rdftoxlsx-apitest",
-            },
-            format="multipart",
-        )
-        self.assertEqual(200, resp.status_code)
-        self.assertEqual(resp.data["message"], "Success")
-        self.assertTrue(resp.data["result"].startswith(settings.MEDIA_URL))
-        self.assertEqual(
-            resp.data["owner"], User.objects.get_by_natural_key(self.username).id
-        )
-        self.client.logout()
-
-    def test_convert_xlsxtordf_api(self):
-        self.client.login(username=self.username, password=self.password)
-        resp = self.client.post(
-            reverse("convert-api"),
-            {
-                "file": self.xlsx_file,
-                "from_format": self.xlsx,
-                "to_format": self.rdf,
-                "cfilename": "xlsxtordf-apitest",
-            },
-            format="multipart",
-        )
-        self.assertEqual(200, resp.status_code)
-        self.assertEqual(resp.data["message"], "Success")
-        self.assertTrue(resp.data["result"].startswith(settings.MEDIA_URL))
-        self.assertEqual(
-            resp.data["owner"], User.objects.get_by_natural_key(self.username).id
-        )
-        self.client.logout()
-
-    def test_convert_xlsxtotag_api(self):
-        self.client.login(username=self.username, password=self.password)
-        resp = self.client.post(
-            reverse("convert-api"),
-            {
-                "file": self.xlsx_file,
-                "from_format": self.xlsx,
-                "to_format": self.tag,
-                "cfilename": "xlsxtotag-apitest",
-            },
-            format="multipart",
-        )
-        self.assertEqual(200, resp.status_code)
-        self.assertEqual(resp.data["message"], "Success")
-        self.assertTrue(resp.data["result"].startswith(settings.MEDIA_URL))
-        self.assertEqual(
-            resp.data["owner"], User.objects.get_by_natural_key(self.username).id
-        )
-        self.client.logout()
-
-    def test_convert_without_one_argument(self):
-        self.client.login(username=self.username, password=self.password)
-        resp = self.client.post(
-            reverse("convert-api"),
-            {
-                "file": self.xlsx_file,
-                "to_format": self.tag,
-                "cfilename": "xlsxtotag-apitest",
-            },
-            format="multipart",
-        )
-        self.assertEqual(resp.status_code, 400)
-
-        resp2 = self.client.post(
-            reverse("convert-api"),
-            {
-                "from_format": self.xlsx,
-                "to_format": self.tag,
-                "cfilename": "xlsxtotag-apitest",
-            },
-            format="multipart",
-        )
-        self.assertEqual(resp2.status_code, 400)
-
-        resp3 = self.client.post(
-            reverse("convert-api"),
-            {"file": self.xlsx_file, "from_format": self.xlsx, "to_format": self.tag},
-            format="multipart",
-        )
-        self.assertEqual(resp3.status_code, 400)
-
-        resp4 = self.client.post(
-            reverse("convert-api"),
-            {
-                "file": self.xlsx_file,
-                "from_format": self.xlsx,
-                "cfilename": "xlsxtotag-apitest",
-            },
-            format="multipart",
-        )
-        self.assertEqual(resp4.status_code, 400)
-        self.client.logout()
-
-
-class CompareFileUploadTests(APITestCase):
-    """Test for compare api with all
-    possible combination of POST and GET
-    request with login enabled.
-    """
-
-    def setUp(self):
-        self.username = "compareapitestuser"
-        self.password = "compareapitestpass"
-        self.tearDown()
-        self.credentials = {"username": self.username, "password": self.password}
-        u = User.objects.create_user(**self.credentials)
-        u.is_staff = True
-        u.save()
-        self.rdf_file = open(getExamplePath("SPDXRdfExample-v2.2.spdx.rdf"))
-        self.rdf_file2 = open(getExamplePath("SPDXRdfExample-v2.3.spdx.rdf"))
-        self.tv_file = open(getExamplePath("SPDXTagExample-v2.0.spdx"))
-
-    def tearDown(self):
-        try:
-            u = User.objects.get_by_natural_key(self.username)
-            u.delete()
-        except ObjectDoesNotExist:
-            pass
-        CompareFileUpload.objects.all().delete()
-
-    def test_compare_api(self):
-        """Access get without login"""
-        resp1 = self.client.get(reverse("compare-api"))
-        self.assertTrue(resp1.status_code, 403)
-        """Access get after login"""
-        self.client.login(username=self.username, password=self.password)
-        resp2 = self.client.get(reverse("compare-api"))
-        self.assertTrue(resp2.status_code, 200)
-        """Compare two valid RDF files"""
-        resp3 = self.client.post(
-            reverse("compare-api"),
-            {
-                "file1": self.rdf_file,
-                "file2": self.rdf_file2,
-                "rfilename": "compare-apitest.xls",
-            },
-            format="multipart",
-        )
-        self.assertEqual(resp3.status_code, 200)
-        self.assertEqual(
-            resp3.data["owner"], User.objects.get_by_natural_key(self.username).id
-        )
-        self.assertTrue(resp3.data["result"].startswith(settings.MEDIA_URL))
-        """Compare with one  invalid RDF files"""
-        resp4 = self.client.post(
-            reverse("compare-api"),
-            {
-                "file1": self.rdf_file,
-                "file2": self.tv_file,
-                "rfilename": "compare-apitest.xls",
-            },
-            format="multipart",
-        )
-        self.assertEqual(resp4.status_code, 400)
-        self.client.logout()
-        self.tearDown()
-
-    def test_compare_without_one_argument(self):
-        self.client.login(username=self.username, password=self.password)
-        resp5 = self.client.post(
-            reverse("compare-api"),
-            {"file1": self.rdf_file, "file2": self.rdf_file2},
-            format="multipart",
-        )
-        self.assertEqual(resp5.status_code, 400)
-
-        resp6 = self.client.post(
-            reverse("compare-api"),
-            {"file1": self.rdf_file, "rfilename": "compare-apitest.xls"},
-            format="multipart",
-        )
-        self.assertEqual(resp6.status_code, 400)
-
-        resp7 = self.client.post(
-            reverse("compare-api"),
-            {"file2": self.rdf_file, "rfilename": "compare-apitest.xls"},
-            format="multipart",
-        )
-        self.assertEqual(resp7.status_code, 400)
-        self.client.logout()
-        self.tearDown()
 
 
 @skipIf(not isRedisAvailable(), "Redis is not available")
@@ -453,11 +66,11 @@ class CheckLicenseFileUploadTests(APITestCase):
     def test_checklicense_api(self):
         """Access get without login"""
         resp1 = self.client.get(reverse("check_license-api"))
-        self.assertTrue(resp1.status_code, 403)
+        self.assertEqual(resp1.status_code, 405)
         self.client.login(username=self.username, password=self.password)
         """ Access get after login"""
         resp2 = self.client.get(reverse("check_license-api"))
-        self.assertTrue(resp2.status_code, 200)
+        self.assertEqual(resp2.status_code, 405)
         """ Valid License File"""
         resp3 = self.client.post(
             reverse("check_license-api"),
@@ -545,7 +158,7 @@ class SubmitLicenseModelsTests(APITestCase):
     def test_submitlicense_api(self):
         """Access submit license api get request without login"""
         resp1 = self.client.get(reverse("submit_license-api"), follow=True, secure=True)
-        self.assertTrue(resp1.status_code, 403)
+        self.assertEqual(resp1.status_code, 200)
         """ Access get after login"""
         app = Application.objects.get(name="Owner")
         token = generate_token()
@@ -559,7 +172,7 @@ class SubmitLicenseModelsTests(APITestCase):
         resp2 = self.client.get(
             reverse("submit_license-api"), HTTP_AUTHORIZATION=self.auth
         )
-        self.assertTrue(resp2.status_code, 201)
+        self.assertEqual(resp2.status_code, 200)
 
     def test_generate_xml(self):
         """View for generating an xml from license submittal form fields"""
@@ -655,7 +268,7 @@ class SubmitLicenseModelsTests(APITestCase):
         resp7 = self.client.post(
             reverse("submit_license-api"), self.data, format="multipart"
         )
-        self.assertTrue(resp7.status_code, 201)
+        self.assertEqual(resp7.status_code, 201)
         self.assertEqual(
             resp7.data["result"],
             "Success! The license request has been successfully submitted.",
